requirements.txt
langchain>=0.3.0
langchain-chroma>=0.1.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
requests>=2.31.0
pydantic>=2.0.0
fastapi>=0.104.0
pandas>=2.0.0
uvicorn>=0.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0